Gtk3: Windows: only copy HTML docs into bindist with --enable-html-docs
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38065 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/make-bindist_win32.sh
@@ -46,6 +46,9 @@ OBJDUMP=$9
 shift
 COMPILER=$9
 
+shift
+HTML_DOCS=$9
+
 
 # Try to get the SVN revision
 #echo "Trying to get SVN revision"
@@ -205,11 +208,15 @@ rm -f `find $BUILDPATH -name "sdl_*"`
 mkdir $BUILDPATH/common
 cp $TOPBUILDDIR/src/arch/gtk3/data/vice.gresource $BUILDPATH/common
 cp $TOPSRCDIR/data/common/C64_Pro_Mono-STYLE.ttf $BUILDPATH/common
-cp -a $TOPSRCDIR/doc/html $BUILDPATH
-cp -a -u $TOPBUILDDIR/doc/html $BUILDPATH
-rm -f $BUILDPATH/html/Makefile* $BUILDPATH/html/checklinks.sh $BUILDPATH/html/texi2html
-rm -f $BUILDPATH/html/robots.txt $BUILDPATH/html/sitemap.xml
-rm -f $BUILDPATH/html/COPYING $BUILDPATH/html/NEWS
+
+if test x"$HTML_DOCS" = "xyes"; then
+    cp -a $TOPSRCDIR/doc/html $BUILDPATH
+    cp -a -u $TOPBUILDDIR/doc/html $BUILDPATH
+    rm -f $BUILDPATH/html/Makefile* $BUILDPATH/html/checklinks.sh $BUILDPATH/html/texi2html
+    rm -f $BUILDPATH/html/robots.txt $BUILDPATH/html/sitemap.xml
+    rm -f $BUILDPATH/html/COPYING $BUILDPATH/html/NEWS
+fi
+
 cp $TOPSRCDIR/COPYING $TOPSRCDIR/FEEDBACK $TOPSRCDIR/NEWS $TOPSRCDIR/README $BUILDPATH
 cp $TOPSRCDIR/doc/readmes/Readme-GTK3.txt $BUILDPATH
 mkdir $BUILDPATH/doc
